@@ -3,8 +3,9 @@ import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { throwIfAlreadyLoaded } from './import-guard/import-guard';
+import { BlobErrorHttpInterceptor } from './http/BlobErrorHttpInterceptor';
 
 import { ErrorDialogComponent } from './messaging/error-dialog/error-dialog.component';
 import { ConfirmDialogComponent } from './messaging/confirm-dialog/confirm-dialog.component';
@@ -26,10 +27,13 @@ import { AngularMaterialModule } from './angular-material/angular-material.modul
 
         HttpClientModule,
     ],
-    entryComponents: [
-        ConfirmDialogComponent,
-        ErrorDialogComponent,
-    ]
+    providers: [
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: BlobErrorHttpInterceptor,
+            multi: true
+        },
+    ],
 })
 export class CoreModule {
     constructor(@Optional() @SkipSelf() parentModule: CoreModule) {
